+++ index.html
@@ -252,6 +252,6 @@
     </div>
   </dialog>
 
-  <script src="app.js"></script>
+  <script type="module" src="src/ui.js"></script>
 </body>
 </html>
